requirements.txt
django==4.2.7
djangorestframework==3.14.0
django_rest_swagger==2.2.0
drf_yasg==1.21.7
isort==5.12.0
django-phonenumber-field==7.2.0
phonenumberslite==8.13.26
djangorestframework_simplejwt==5.3.0